requirements.txt
Flask
joblib
pandas
numpy
matplotlib
mysql-connector-python
scikit-learn
xgboost
gunicorn
python-dotenv